requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
--extra-index-url https://pypi.ngc.nvidia.com
--trusted-host pypi.ngc.nvidia.com

aeon==0.4.0
    # via -r requirements.in
attrs~=24.3.0
    # via aeon
deprecated==1.2.14
    # via aeon
filelock==3.12.4
    # via torch
jinja2~=3.1.6
    # via torch
joblib==1.3.2
    # via scikit-learn
llvmlite==0.40.1
    # via numba
markupsafe~=3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.1
    # via torch
numba~=0.60.0
    # via aeon
numpy~=1.26.4
    # via
    #   -r requirements.in
    #   aeon
    #   numba
    #   pandas
    #   scikit-learn
    #   scipy
packaging~=24.2
    # via aeon
pandas~=2.2.3
    # via
    #   -r requirements.in
    #   aeon
python-dateutil~=2.9.0post0
    # via pandas
pytz~=2024.1
    # via pandas
scikit-learn~=1.3.0
    # via
    #   -r requirements.in
    #   aeon
scipy~=1.12.0
    # via
    #   aeon
    #   scikit-learn
six==1.16.0
    # via python-dateutil
sympy~=1.13.3
    # via torch
threadpoolctl==3.2.0
    # via scikit-learn
torch~=2.2.2
    # via -r requirements.in
typing-extensions==4.8.0
    # via torch
tzdata~=2025a
    # via pandas
wrapt==1.15.0
    # via deprecated

setuptools~=75.8.0

seaborn~=0.13.2

shap~=0.46.0
xrocket~=0.2.1
GitPython~=3.1.43
gitdb~=4.0.7
cffi~=1.17.1
defusedxml~=0.7.1
Pillow~=11.1.0
ipython~=8.15.0
pip~=25.0
wheel~=0.45.1
protobuf~=4.25.3
jupyter~=1.0.0
jupyterlab-pygments~=0.2.2
rich~=13.9.4
Pygments~=2.15.1
markdown-it-py~=2.2.0
toml~=0.10.2
zipp~=3.21.0
smmap~=4.0.0
mdurl~=0.1.0
charset-normalizer~=3.3.2
altair~=5.0.1
jsonschema~=4.23.0
narwhals~=1.29.0
pyarrow~=19.0.0
traitlets~=5.14.3
PyYAML~=6.0
matplotlib~=3.6.3
pydeck~=0.9.1
ipywidgets~=8.0.4
certifi~=2025.1.31
tornado~=6.2
urllib3~=2.3.0
requests~=2.32.3
idna~=3.4
click~=8.1.7
streamlit~=1.42.0
cachetools~=5.5.1
blinker~=1.9.0
webcolors~=1.12
jsonpointer~=2.3
uri-template~=1.2.0
isoduration~=20.11.0
fqdn~=1.5.1
rfc3339-validator~=0.1.4
typing_extensions~=4.12.2
zlib~=1.2.13
brotli~=1.0.9
pysocks~=1.7.1
openssl~=3.0.16
overrides~=7.4.0
referencing~=0.30.2
yaml~=0.2.5
numexpr~=2.10.1
tenacity~=9.0.0
jsonschema-specifications~=2023.7.1
importlib_resources~=6.4.0
reportlab~=3.6.13
sktime~=0.36.0
lime~=0.2.0.1